AS7-6120 Add transformation to threads resources Fix expression rejection test infrastructure
