+++ model-test/src/main/java/org/jboss/as/model/test/FailedOperationTransformationConfig.java
@@ -347,6 +347,7 @@ public class FailedOperationTransformationConfig {
                         return true;
                     }
                 }
+                break;
             case PROPERTY:
                 if (hasExpressions(attribute.asProperty().getValue())) {
                     return true;
