+++ threads/src/main/java/org/jboss/as/threads/QueuelessThreadPoolResourceDefinition.java
@@ -22,12 +22,22 @@
 
 package org.jboss.as.threads;
 
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.WRITE_ATTRIBUTE_OPERATION;
+
+import java.util.HashMap;
+import java.util.Map;
+
 import org.jboss.as.controller.OperationStepHandler;
 import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.ReadResourceNameOperationStepHandler;
 import org.jboss.as.controller.ResourceDefinition;
 import org.jboss.as.controller.SimpleResourceDefinition;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
+import org.jboss.as.controller.transform.AttributeTransformationRequirementChecker;
+import org.jboss.as.controller.transform.RejectExpressionValuesTransformer;
+import org.jboss.as.controller.transform.ResourceTransformer;
+import org.jboss.as.controller.transform.TransformersSubRegistration;
 import org.jboss.msc.service.ServiceName;
 
 /**
@@ -91,5 +101,21 @@ public class QueuelessThreadPoolResourceDefinition extends SimpleResourceDefinit
         if (registerRuntimeOnly) {
             new QueuelessThreadPoolMetricsHandler(serviceNameBase).registerAttributes(resourceRegistration);
         }
+    }
+
+    public static void registerTransformers1_0(TransformersSubRegistration parent) {
+        registerTransformers1_0(parent, CommonAttributes.BLOCKING_QUEUELESS_THREAD_POOL);
+        registerTransformers1_0(parent, CommonAttributes.QUEUELESS_THREAD_POOL);
+    }
+
+    public static void registerTransformers1_0(TransformersSubRegistration parent, String type) {
+
+        final RejectExpressionValuesTransformer TRANSFORMER =
+                new RejectExpressionValuesTransformer(PoolAttributeDefinitions.KEEPALIVE_TIME.getName(),
+                                                    KeepAliveTimeAttributeDefinition.TRANSFORMATION_REQUIREMENT_CHECKER);
+
+        final TransformersSubRegistration pool = parent.registerSubResource(PathElement.pathElement(type), (ResourceTransformer) TRANSFORMER);
+        pool.registerOperationTransformer(ADD, TRANSFORMER);
+        pool.registerOperationTransformer(WRITE_ATTRIBUTE_OPERATION, TRANSFORMER.getWriteAttributeTransformer());
     }
 }
