+++ threads/src/main/java/org/jboss/as/threads/ThreadFactoryResourceDefinition.java
@@ -22,12 +22,18 @@
 
 package org.jboss.as.threads;
 
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.WRITE_ATTRIBUTE_OPERATION;
+
 import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.ReadResourceNameOperationStepHandler;
 import org.jboss.as.controller.ResourceDefinition;
 import org.jboss.as.controller.SimpleResourceDefinition;
 import org.jboss.as.controller.descriptions.StandardResourceDescriptionResolver;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
+import org.jboss.as.controller.transform.RejectExpressionValuesTransformer;
+import org.jboss.as.controller.transform.ResourceTransformer;
+import org.jboss.as.controller.transform.TransformersSubRegistration;
 
 /**
  * {@link ResourceDefinition} for a thread factory resource.
@@ -52,5 +58,19 @@ public class ThreadFactoryResourceDefinition extends SimpleResourceDefinition {
     public void registerAttributes(ManagementResourceRegistration resourceRegistration) {
         resourceRegistration.registerReadOnlyAttribute(PoolAttributeDefinitions.NAME, ReadResourceNameOperationStepHandler.INSTANCE);
         ThreadFactoryWriteAttributeHandler.INSTANCE.registerAttributes(resourceRegistration);
+    }
+
+    public static void registerTransformers1_0(TransformersSubRegistration parent) {
+        registerTransformers1_0(parent, CommonAttributes.THREAD_FACTORY);
+    }
+
+    public static void registerTransformers1_0(TransformersSubRegistration parent, String type) {
+
+        final RejectExpressionValuesTransformer TRANSFORMER =
+                new RejectExpressionValuesTransformer(PoolAttributeDefinitions.GROUP_NAME, PoolAttributeDefinitions.THREAD_NAME_PATTERN);
+
+        final TransformersSubRegistration pool = parent.registerSubResource(PathElement.pathElement(type), (ResourceTransformer) TRANSFORMER);
+        pool.registerOperationTransformer(ADD, TRANSFORMER);
+        pool.registerOperationTransformer(WRITE_ATTRIBUTE_OPERATION, TRANSFORMER.getWriteAttributeTransformer());
     }
 }
