+++ threads/src/main/java/org/jboss/as/threads/ThreadsExtension.java
@@ -21,18 +21,21 @@
  */
 package org.jboss.as.threads;
 
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.DESCRIBE;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.SUBSYSTEM;
 import static org.jboss.as.threads.CommonAttributes.THREADS;
 
 import org.jboss.as.controller.Extension;
 import org.jboss.as.controller.ExtensionContext;
+import org.jboss.as.controller.ModelVersion;
+import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.SubsystemRegistration;
 import org.jboss.as.controller.descriptions.ResourceDescriptionResolver;
 import org.jboss.as.controller.descriptions.StandardResourceDescriptionResolver;
 import org.jboss.as.controller.operations.common.GenericSubsystemDescribeHandler;
 import org.jboss.as.controller.parsing.ExtensionParsingContext;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
-import org.jboss.as.controller.registry.OperationEntry;
+import org.jboss.as.controller.transform.ResourceTransformer;
+import org.jboss.as.controller.transform.TransformersSubRegistration;
 
 /**
  * Extension for thread management.
@@ -43,11 +46,12 @@ import org.jboss.as.controller.registry.OperationEntry;
 public class ThreadsExtension implements Extension {
 
     public static String SUBSYSTEM_NAME = "threads";
+    static final PathElement SUBSYSTEM_PATH = PathElement.pathElement(SUBSYSTEM, SUBSYSTEM_NAME);
 
     static final String RESOURCE_NAME = ThreadsExtension.class.getPackage().getName() + ".LocalDescriptions";
 
     private static final int MANAGEMENT_API_MAJOR_VERSION = 1;
-    private static final int MANAGEMENT_API_MINOR_VERSION = 0;
+    private static final int MANAGEMENT_API_MINOR_VERSION = 1;
     private static final int MANAGEMENT_API_MICRO_VERSION = 0;
 
     public static ResourceDescriptionResolver getResourceDescriptionResolver(final String keyPrefix, boolean useUnprefixedChildTypes) {
@@ -69,6 +73,8 @@ public class ThreadsExtension implements Extension {
         // Remoting subsystem description and operation handlers
         final ManagementResourceRegistration subsystem = registration.registerSubsystemModel(new ThreadSubsystemResourceDefinition(registerRuntimeOnly));
         subsystem.registerOperationHandler(GenericSubsystemDescribeHandler.DEFINITION, GenericSubsystemDescribeHandler.INSTANCE);
+
+        registerTransformers1_0(registration);
     }
 
     @Override
@@ -76,5 +82,25 @@ public class ThreadsExtension implements Extension {
         context.setSubsystemXmlMapping(SUBSYSTEM_NAME, Namespace.CURRENT.getUriString(), ThreadsParser.INSTANCE);
         context.setSubsystemXmlMapping(SUBSYSTEM_NAME, Namespace.THREADS_1_0.getUriString(), ThreadsParser.INSTANCE);
     }
+
+    // Transformation
+
+    /**
+     * Register the transformers for older model versions.
+     *
+     * @param subsystem the subsystems registration
+     */
+    private static void registerTransformers1_0(final SubsystemRegistration subsystem) {
+
+        // Transformation to model version 1.0.0
+        final ModelVersion version100 = ModelVersion.create(1, 0, 0);
+        final TransformersSubRegistration registration = subsystem.registerModelTransformers(version100, ResourceTransformer.DEFAULT);
+        BoundedQueueThreadPoolResourceDefinition.registerTransformers1_0(registration);
+        QueuelessThreadPoolResourceDefinition.registerTransformers1_0(registration);
+        ScheduledThreadPoolResourceDefinition.registerTransformers1_0(registration);
+        UnboundedQueueThreadPoolResourceDefinition.registerTransformers1_0(registration);
+        ThreadFactoryResourceDefinition.registerTransformers1_0(registration);
+    }
+
 
 }
