+++ threads/src/main/java/org/jboss/as/threads/UnboundedQueueThreadPoolResourceDefinition.java
@@ -22,11 +22,17 @@
 
 package org.jboss.as.threads;
 
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD;
+import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.WRITE_ATTRIBUTE_OPERATION;
+
 import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.ReadResourceNameOperationStepHandler;
 import org.jboss.as.controller.ResourceDefinition;
 import org.jboss.as.controller.SimpleResourceDefinition;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
+import org.jboss.as.controller.transform.RejectExpressionValuesTransformer;
+import org.jboss.as.controller.transform.ResourceTransformer;
+import org.jboss.as.controller.transform.TransformersSubRegistration;
 import org.jboss.msc.service.ServiceName;
 
 /**
@@ -66,5 +72,20 @@ public class UnboundedQueueThreadPoolResourceDefinition extends SimpleResourceDe
         if (registerRuntimeOnly) {
             new UnboundedQueueThreadPoolMetricsHandler(serviceNameBase).registerAttributes(resourceRegistration);
         }
+    }
+
+    public static void registerTransformers1_0(TransformersSubRegistration parent) {
+        registerTransformers1_0(parent, CommonAttributes.UNBOUNDED_QUEUE_THREAD_POOL);
+    }
+
+    public static void registerTransformers1_0(TransformersSubRegistration parent, String type) {
+
+        final RejectExpressionValuesTransformer TRANSFORMER =
+                new RejectExpressionValuesTransformer(PoolAttributeDefinitions.KEEPALIVE_TIME.getName(),
+                        KeepAliveTimeAttributeDefinition.TRANSFORMATION_REQUIREMENT_CHECKER);
+
+        final TransformersSubRegistration pool = parent.registerSubResource(PathElement.pathElement(type), (ResourceTransformer) TRANSFORMER);
+        pool.registerOperationTransformer(ADD, TRANSFORMER);
+        pool.registerOperationTransformer(WRITE_ATTRIBUTE_OPERATION, TRANSFORMER.getWriteAttributeTransformer());
     }
 }
