--- threads/src/test/java/org/jboss/as/threads/ThreadsSubsystem11Test.java
@@ -1,49 +1,0 @@
-/*
- *
- *  JBoss, Home of Professional Open Source.
- *  Copyright 2012, Red Hat, Inc., and individual contributors
- *  as indicated by the @author tags. See the copyright.txt file in the
- *  distribution for a full listing of individual contributors.
- *
- *  This is free software; you can redistribute it and/or modify it
- *  under the terms of the GNU Lesser General Public License as
- *  published by the Free Software Foundation; either version 2.1 of
- *  the License, or (at your option) any later version.
- *
- *  This software is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- *  Lesser General Public License for more details.
- *
- *  You should have received a copy of the GNU Lesser General Public
- *  License along with this software; if not, write to the Free
- *  Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- *  02110-1301 USA, or see the FSF site: http://www.fsf.org.
- * /
- */
-
-package org.jboss.as.threads;
-
-import java.io.IOException;
-
-import org.jboss.as.subsystem.test.AbstractSubsystemBaseTest;
-import org.junit.Test;
-
-/**
- * @author <a href="mailto:tomaz.cerar@redhat.com">Tomaz Cerar</a> (c) 2012 Red Hat Inc.
- */
-public class ThreadsSubsystem11Test extends AbstractSubsystemBaseTest {
-    public ThreadsSubsystem11Test() {
-        super(ThreadsExtension.SUBSYSTEM_NAME, new ThreadsExtension());
-    }
-
-    @Override
-    protected String getSubsystemXml() throws IOException {
-        return readResource("threads-subsystem-1_1.xml");
-    }
-
-    @Test
-    public void testExpressions() throws Exception {
-        standardSubsystemTest("expressions.xml");
-    }
-}
