+++ threads/src/test/resources/org/jboss/as/threads/threads-transform-1_0.xml
@@ -1,0 +1,42 @@
+<subsystem xmlns="urn:jboss:domain:threads:1.1">
+    <thread-factory name="test-factory" group-name="Thread Group" thread-name-pattern="%G - %t" priority="${test.exp:6}"/>
+    <unbounded-queue-thread-pool name="test-pool">
+        <max-threads count="${prop.max-thread-count:100}"/>
+        <keepalive-time time="1000" unit="milliseconds"/>
+        <thread-factory name="test-factory"/>
+    </unbounded-queue-thread-pool>
+    <bounded-queue-thread-pool name="test-pool" allow-core-timeout="true">
+        <core-threads count="${test.exp:200}"/>
+        <queue-length count="${test.exp:300}"/>
+        <max-threads count="${test.exp:100}"/>
+        <keepalive-time time="1000" unit="milliseconds"/>
+        <thread-factory name="test-factory"/>
+        <handoff-executor name="other"/>
+    </bounded-queue-thread-pool>
+    <blocking-bounded-queue-thread-pool name="test-pool" allow-core-timeout="true">
+        <core-threads count="${test.exp:200}"/>
+        <queue-length count="${test.exp:300}"/>
+        <max-threads count="${test.exp:100}"/>
+        <keepalive-time time="1000" unit="milliseconds"/>
+        <thread-factory name="test-factory"/>
+    </blocking-bounded-queue-thread-pool>
+    <queueless-thread-pool name="other">
+        <max-threads count="${test.exp:1}"/>
+    </queueless-thread-pool>
+    <queueless-thread-pool name="test-pool">
+        <max-threads count="${prop.max-thread-count:100}"/>
+        <keepalive-time time="1000" unit="milliseconds"/>
+        <thread-factory name="test-factory"/>
+        <handoff-executor name="other"/>
+    </queueless-thread-pool>
+    <blocking-queueless-thread-pool name="test-pool">
+        <max-threads count="${prop.max-thread-count:100}"/>
+        <keepalive-time time="1000" unit="milliseconds"/>
+        <thread-factory name="test-factory"/>
+    </blocking-queueless-thread-pool>
+    <scheduled-thread-pool name="test-pool">
+        <max-threads count="${prop.max-thread-count:10}"/>
+        <keepalive-time time="1000" unit="milliseconds"/>
+        <thread-factory name="test-factory"/>
+    </scheduled-thread-pool>
+</subsystem>
